+++ src/javascript/esutil.ts
@@ -16,7 +16,7 @@ import generate from 'babel-generator';
 import * as babel from 'babel-types';
 import * as doctrine from 'doctrine';
 
-import {MethodParam, ScannedMethod} from '../index';
+import {MethodParam, ScannedMethod, ScannedProperty} from '../index';
 import {Result} from '../model/analysis';
 import {ImmutableSet} from '../model/immutable';
 import {Privacy} from '../model/model';
@@ -427,4 +427,167 @@ function* _getMethods(node: babel.Node) {
       yield statement;
     }
   }
+}
+
+/*
+ * Extracts a property from a given getter or setter method,
+ * whether it be an object method or a class method.
+ */
+export function extractPropertyFromGetterOrSetter(
+    method: babel.ClassMethod|babel.ObjectMethod,
+    jsdocAnn: jsdoc.Annotation|undefined,
+    document: JavaScriptDocument): ScannedProperty|null {
+  // TODO(43081j): remove this when static properties are supported
+  if (babel.isClassMethod(method) && method.static) {
+    return null;
+  }
+
+  if (method.kind !== 'get' && method.kind !== 'set') {
+    return null;
+  }
+
+  // TODO(43081j): use getPropertyName, see
+  // https://github.com/Polymer/polymer-analyzer/pull/867
+  if (!babel.isIdentifier(method.key)) {
+    return null;
+  }
+
+  if (method.key.name === undefined) {
+    return null;
+  }
+
+  let type;
+  let description;
+  let privacy: Privacy = 'public';
+  let readOnly = false;
+
+  if (jsdocAnn) {
+    const ret = getReturnFromAnnotation(jsdocAnn);
+    type = ret ? ret.type : undefined;
+    description = jsdoc.getDescription(jsdocAnn);
+    privacy = getOrInferPrivacy(method.key.name, jsdocAnn);
+    readOnly = jsdoc.hasTag(jsdocAnn, 'readonly');
+  }
+
+  return {
+    name: method.key.name,
+    astNode: method,
+    type,
+    jsdoc: jsdocAnn,
+    sourceRange: document.sourceRangeForNode(method)!,
+    description,
+    privacy,
+    warnings: [],
+    readOnly,
+  };
+}
+
+/**
+ * Extracts properties (including accessors) from a given class
+ * or object expression.
+ */
+export function extractPropertiesFromClassOrObjectBody(
+    node: babel.Class|babel.ObjectExpression,
+    document: JavaScriptDocument): Map<string, ScannedProperty> {
+  const properties = new Map<string, ScannedProperty>();
+  const accessors = new Map<
+      string,
+      {getter?: babel.ClassMethod|babel.ObjectMethod, setter?: babel.ClassMethod|babel.ObjectMethod}>();
+
+  let body;
+
+  if (babel.isClass(node)) {
+    body = node.body.body;
+  } else {
+    body = node.properties;
+  }
+
+  for (const member of body) {
+    if (babel.isSpreadProperty(member)) {
+      continue;
+    }
+
+    if ((babel.isMethod(member) || babel.isObjectProperty(member)) &&
+        member.computed) {
+      continue;
+    }
+
+    const name = astValue.getIdentifierName(member.key)!;
+
+    if (babel.isMethod(member) ||
+        babel.isFunction(member.value)) {
+      if (babel.isMethod(member) &&
+          (member.kind === 'get' || member.kind === 'set')) {
+        let accessor = accessors.get(name);
+
+        if (!accessor) {
+          accessor = {};
+          accessors.set(name, accessor);
+        }
+
+        if (member.kind === 'get') {
+          accessor.getter = member;
+        } else {
+          accessor.setter = member;
+        }
+      }
+
+      continue;
+    }
+
+    const astNode = member.key;
+    const sourceRange = document.sourceRangeForNode(member)!;
+    const jsdocAnn = jsdoc.parseJsdoc(getAttachedComment(member) || '');
+    const detectedType =
+        getClosureType(member.value, jsdocAnn, sourceRange, document);
+    let type: string|undefined = undefined;
+
+    if (detectedType.successful) {
+      type = detectedType.value;
+    }
+
+    properties.set(name, {
+      name,
+      astNode,
+      type,
+      jsdoc: jsdocAnn,
+      sourceRange,
+      description: jsdocAnn ? jsdoc.getDescription(jsdocAnn) : undefined,
+      privacy: getOrInferPrivacy(name, jsdocAnn),
+      warnings: [],
+      readOnly: jsdoc.hasTag(jsdocAnn, 'readonly'),
+    });
+  }
+
+  for (const val of accessors.values()) {
+    let getter: ScannedProperty|null = null;
+    let setter: ScannedProperty|null = null;
+
+    if (val.getter) {
+      const parsedJsdoc =
+          jsdoc.parseJsdoc(getAttachedComment(val.getter) || '');
+      getter = extractPropertyFromGetterOrSetter(
+          val.getter, parsedJsdoc, document);
+    }
+
+    if (val.setter) {
+      const parsedJsdoc =
+          jsdoc.parseJsdoc(getAttachedComment(val.setter) || '');
+      setter = extractPropertyFromGetterOrSetter(
+          val.setter, parsedJsdoc, document);
+    }
+
+    const prop = getter || setter;
+    if (!prop) {
+      continue;
+    }
+
+    if (!prop.readOnly) {
+      prop.readOnly = (val.setter === undefined);
+    }
+
+    properties.set(prop.name, prop);
+  }
+
+  return properties;
 }